Renamed `src/core/Wizard` to `src/core/Stepper`
