+++ packages/iTwinUI-react/src/core/Stepper/StepperStep.tsx
